unify fetching of all properties
