+++ src/routes/category-properties/+page.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
 	import PropertyListPage from '$pages/PropertyListPage.svelte'
+
 	let { data } = $props()
 </script>
