+++ src/routes/category-search/+page.svelte
@@ -6,7 +6,7 @@
 	const sample_search_url = `/category-search/results?satisfied=finitely_complete~pointed&unsatisfied=complete`
 </script>
 
-<SearchPage type="category" all_properties={data.all_properties}>
+<SearchPage type="category" {...data}>
 	Search for categories with certain properties while excluding others. For example, you
 	can <a href={sample_search_url}>look</a> for categories that are finitely complete and pointed
 	but not complete.
