+++ src/routes/functor-search/+page.server.ts
@@ -1,4 +1,4 @@
-import { get_property_ids } from '$lib/server/properties'
+import { get_property_ids } from '$lib/server/fetchers/properties'
 
 export const load = () => {
 	return { all_properties: get_property_ids('functor') }
