+++ src/routes/functor-search/+page.svelte
@@ -6,7 +6,7 @@
 	const sample_search_url = `/functor-search/results?satisfied=continuous&unsatisfied=cocontinuous`
 </script>
 
-<SearchPage all_properties={data.all_properties} type="functor">
+<SearchPage type="functor" {...data}>
 	Search for functors with certain properties while excluding others. For example, you
 	can <a href={sample_search_url}>look</a> for functors that are continuous but not cocontinuous.
 </SearchPage>
